Moved file-based storage interface to separate file
