+++ plaso/storage/sqlite/reader.py
@@ -3,11 +3,11 @@
 
 from __future__ import unicode_literals
 
-from plaso.storage import interface
+from plaso.storage import file_interface
 from plaso.storage.sqlite import sqlite_file
 
 
-class SQLiteStorageFileReader(interface.StorageFileReader):
+class SQLiteStorageFileReader(file_interface.StorageFileReader):
   """SQLite-based storage file reader."""
 
   def __init__(self, path):
